Add missing metadata to search page, improve pagination description
